// src/contracts/generated/csi10/BrowserLogoutCleanupPageAcknowledgementV1.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { Secret32 } from "./Secret32";

export type BrowserLogoutCleanupPageAcknowledgementV1 = { schemaVersion: number, operationId: string, logoutCsrfToken: Secret32, pageId: string, };