// src/contracts/generated/csi10/BrowserLogoutCommitRequestV1.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BrowserHeadReferenceV1 } from "./BrowserHeadReferenceV1";
import type { Secret32 } from "./Secret32";

export type BrowserLogoutCommitRequestV1 = { schemaVersion: number, operationId: string, logoutCsrfToken: Secret32, head: BrowserHeadReferenceV1, };