// src/contracts/generated/csi10/BrowserLogoutOptionsResultV1.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { LogoutAccountReferenceV1 } from "./LogoutAccountReferenceV1";

export type BrowserLogoutOptionsResultV1 = { schemaVersion: number, accounts: Array<LogoutAccountReferenceV1>, expiresAt: string, };