// src/contracts/generated/csi10/BrowserLogoutStatusRequestV1.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BrowserLogoutStatusProofV1 } from "./BrowserLogoutStatusProofV1";

export type BrowserLogoutStatusRequestV1 = { schemaVersion: number, operationId: string, proof: BrowserLogoutStatusProofV1, };