// src/contracts/generated/csi10/LogoutAccountMetadataResultV1.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AvatarV1 } from "./AvatarV1";
import type { BrowserAccountId } from "./BrowserAccountId";
import type { HomeTenantTypeV1 } from "./HomeTenantTypeV1";

export type LogoutAccountMetadataResultV1 = { "kind": "current", schemaVersion: number, browserAccountId: BrowserAccountId, displayName: string, primaryEmail: string, avatar: AvatarV1, homeTenantLabel: string, homeTenantType: HomeTenantTypeV1, sessionExpiresAt: string, } | { "kind": "invalid", schemaVersion: number, browserAccountId: BrowserAccountId, };