// src/contracts/generated/csi10/LogoutAccountReferenceV1.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { BrowserAccountId } from "./BrowserAccountId";
import type { CatalogId } from "./CatalogId";
import type { Origin } from "./Origin";

export type LogoutAccountReferenceV1 = { browserAccountId: BrowserAccountId, homeRegionId: CatalogId, identityApiOrigin: Origin, metadataCapability: string, expiresAt: string, };